requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
beautifulsoup4==4.14.3
    # via deep-translator
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==46.0.7
    # via google-auth
deep-translator==1.11.4
    # via ai-doc-generator (pyproject.toml)
dj-database-url==3.1.2
    # via ai-doc-generator (pyproject.toml)
django==5.0.14
    # via
    #   ai-doc-generator (pyproject.toml)
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via ai-doc-generator (pyproject.toml)
djangorestframework==3.17.1
    # via ai-doc-generator (pyproject.toml)
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.30.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.193.0
    # via google-generativeai
google-auth==2.49.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.1
    # via google-api-python-client
google-generativeai==0.8.6
    # via ai-doc-generator (pyproject.toml)
googleapis-common-protos==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.80.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==25.3.0
    # via ai-doc-generator (pyproject.toml)
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
packaging==26.0
    # via gunicorn
pillow==12.2.0
    # via ai-doc-generator (pyproject.toml)
proto-plus==1.27.2
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via ai-doc-generator (pyproject.toml)
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
python-dotenv==1.2.2
    # via ai-doc-generator (pyproject.toml)
requests==2.33.1
    # via
    #   deep-translator
    #   google-api-core
soupsieve==2.8.3
    # via beautifulsoup4
sqlparse==0.5.5
    # via django
tqdm==4.67.3
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2026.1
    # via django
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
whitenoise==6.12.0
    # via ai-doc-generator (pyproject.toml)
